custom-auth/requirements.txt
python-jose==3.2.0